+++ custom_components/vag_connect/config_flow.py
@@ -119,6 +119,7 @@ class VagConnectConfigFlow(config_entries.ConfigFlow, domain=DOMAIN):
                     vol.Optional(
                         CONF_SCAN_INTERVAL, default=DEFAULT_SCAN_INTERVAL
                     ): vol.All(vol.Coerce(int), vol.Range(min=MIN_SCAN_INTERVAL)),
+                    vol.Optional("force_enable_access", default=False): bool,
                 }
             ),
             errors=errors,